HAP-281 fix tests

+++ src/sh/run_quantify_test.sh
@@ -15,6 +15,8 @@ HG19=${DIR}/../../example/chr21.fa
 TMP_OUT=`mktemp -t happy.XXXXXXXXXX`
 
 # run hap.py on NA12878 test file
+# note this only works if we don't change the confident regions
+# based on the truthset
 ${PYTHON} ${HCDIR}/hap.py \
 			 	-l chr21 \
 			 	${DIR}/../../example/happy/PG_NA12878_chr21.vcf.gz \
@@ -22,7 +24,7 @@ ${PYTHON} ${HCDIR}/hap.py \
 			 	-f ${DIR}/../../example/happy/PG_Conf_chr21.bed.gz \
 			 	-r ${DIR}/../../example/chr21.fa \
 			 	-o ${TMP_OUT} \
-			 	-V \
+                --no-adjust-conf-regions \
 			 	--force-interactive
 
 if [[ $? != 0 ]]; then
@@ -40,11 +42,13 @@ fi
 # hap.py writes GA4GH-compliant output VCFs
 # we should be able to re-quantify using the GA4GH
 # spec and get the same result
+# Note that the default is now to run with --adjust-conf-regions in hap.py,
+# so we have to also pass the truth VCF here
 ${PYTHON} ${HCDIR}/qfy.py \
            ${TMP_OUT}.vcf.gz \
            -o ${TMP_OUT}.qfy  \
            -f ${DIR}/../../example/happy/PG_Conf_chr21.bed.gz \
-           -t ga4gh -X -V
+           -t ga4gh -X --verbose
 
 if [[ $? != 0 ]]; then
 	echo "qfy.py failed!"
@@ -62,11 +66,6 @@ if [[ $? != 0 ]] || [[ ! -s ${TMP_OUT}.qfy.m.json ]]; then
 	echo "Cannot unzip metrics for re-quantified run."
 	exit 1
 fi
-diff ${TMP_OUT}.hap.m.json ${TMP_OUT}.qfy.m.json
-if [[ $? != 0 ]]; then
-	echo "Re-quantified counts are different! diff ${TMP_OUT}.hap.m.json ${TMP_OUT}.qfy.m.json "
-	exit 1
-fi
 
 # This script checks if the summary precision / recall figures have changed significantly
 ${PYTHON} ${DIR}/compare_summaries.py ${TMP_OUT}.qfy.summary.csv ${DIR}/../../example/happy/expected.summary.csv
@@ -74,5 +73,12 @@ if [[ $? != 0 ]]; then
 	echo "All summary differs! -- diff ${TMP_OUT}.qfy.summary.csv ${DIR}/../../example/happy/expected.summary.csv"
 	exit 1
 fi
+
+diff ${TMP_OUT}.hap.m.json ${TMP_OUT}.qfy.m.json > /dev/null
+if [[ $? != 0 ]]; then
+	echo "Re-quantified counts are different! diff ${TMP_OUT}.hap.m.json ${TMP_OUT}.qfy.m.json "
+	exit 1
+fi
+
 
 rm -rf ${TMP_OUT}.*
